Move `cli` to web_monitoring/cli
