+++ scripts/wm
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from web_monitoring.cli import main
+from web_monitoring.cli.cli import main
 
 
 if __name__ == '__main__':
